requirements.txt
# List of all Python packages to install
requests
protobuf
gtfs-realtime-bindings
python-dotenv